requirements.txt
crewai
crewai-tools
langchain-community
langchain
langchain-huggingface
langchain-chroma
sentence-transformers
python-dotenv
tavily-python
flask==3.0.0
flask-cors==4.0.0
gunicorn==21.2.0
werkzeug==3.0.0
